Move hooks into reminders
