+++ client/src/lib/reminders/hooks/usePracticeReminderNotificationsSetting.ts
@@ -1,8 +1,8 @@
 import {useCallback, useEffect, useMemo, useState} from 'react';
 import useCurrentUserState from '../../user/hooks/useCurrentUserState';
 import useUserState, {PracticeReminderConfig} from '../../user/state/state';
-import useNotificationPermissions from './useNotificationPermissions';
-import useUpdatePracticeNotifications from '../../reminders/useUpdatePracticeNotifications';
+import useNotificationPermissions from '../../notifications/hooks/useNotificationPermissions';
+import useUpdatePracticeNotifications from './useUpdatePracticeNotifications';
 
 const usePracticeReminderNotificationsSetting = () => {
   const userState = useCurrentUserState();
