+++ client/src/lib/reminders/hooks/useUpdatePracticeNotifications.ts
@@ -2,19 +2,19 @@ import {useCallback} from 'react';
 import dayjs from 'dayjs';
 import utc from 'dayjs/plugin/utc';
 
-import {NOTIFICATION_CHANNELS} from '../notifications/constants';
-import usePinnedCollections from '../user/hooks/usePinnedCollections';
-import useGetCollectionById from '../content/hooks/useGetCollectionById';
-import useUserEvents from '../user/hooks/useUserEvents';
-import useTriggerNotifications from '../notifications/hooks/useTriggerNotifications';
-import {calculateNextReminderTime} from './utils';
+import {NOTIFICATION_CHANNELS} from '../../notifications/constants';
+import usePinnedCollections from '../../user/hooks/usePinnedCollections';
+import useGetCollectionById from '../../content/hooks/useGetCollectionById';
+import useUserEvents from '../../user/hooks/useUserEvents';
+import useTriggerNotifications from '../../notifications/hooks/useTriggerNotifications';
+import {calculateNextReminderTime} from '../utils';
 import {useTranslation} from 'react-i18next';
-import {Collection} from '../../../../shared/src/types/generated/Collection';
-import {PracticeReminderConfig} from '../user/state/state';
+import {Collection} from '../../../../../shared/src/types/generated/Collection';
+import {PracticeReminderConfig} from '../../user/state/state';
 import {
   DEFAULT_NUMBER_OF_PRACTICE_REMINDERS,
   REMINDER_INTERVALS,
-} from './constants';
+} from '../constants';
 
 dayjs.extend(utc);
